+++ CHANGELOG.rst
@@ -2,9 +2,11 @@
 Changelog
 =========
 
-TBR
+
+TBR:
 ------------------
 
+* Cache mechanism using SCRAPY_POET_CACHE setting
 * removed support for Python 3.6
 * added support for Python 3.10